pipline2/extract/requirements.txt
paho-mqtt
kafka-python
opencv-python
numpy
pillow